+++ python/jet/interpreter.py
@@ -1,14 +1,17 @@
+import random
+import warnings
 from copy import deepcopy
 from inspect import signature
 from typing import Any, Callable, Dict, Iterator, List, Sequence, Set, Union
-from warnings import warn
 
 import numpy as np
 
 from xir import Statement, XIRProgram
 from xir import parse_script as parse_xir_script
 
+from .bindings import PathInfo
 from .circuit import Circuit
+from .factory import TaskBasedContractor, TensorNetworkType, TensorType
 from .gate import GateFactory
 from .state import Qudit
 
@@ -127,7 +130,7 @@ def run_xir_program(program: XIRProgram) -> List[Union[np.number, np.ndarray]]:
             gate = GateFactory.create(stmt.name, **stmt.params)
             circuit.append_gate(gate, wire_ids=stmt.wires)
 
-        elif stmt.name in ("amplitude", "Amplitude"):
+        elif stmt.name in ("Amplitude", "amplitude"):
             if not isinstance(stmt.params, dict) or "state" not in stmt.params:
                 raise ValueError(f"Statement '{stmt}' is missing a 'state' parameter.")
 
@@ -154,6 +157,13 @@ def run_xir_program(program: XIRProgram) -> List[Union[np.number, np.ndarray]]:
                 raise ValueError(f"Statement '{stmt}' must be applied to [0 .. {num_wires - 1}].")
 
             output = _compute_amplitude(circuit=circuit, state=state)
+            result.append(output)
+
+        elif stmt.name in ("Probabilities", "probabilities"):
+            if stmt.wires != tuple(range(num_wires)):
+                raise ValueError(f"Statement '{stmt}' must be applied to [0 .. {num_wires - 1}].")
+
+            output = _compute_probabilities(circuit=circuit)
             result.append(output)
 
         else:
@@ -184,7 +194,7 @@ def _resolve_xir_program_statements(program: XIRProgram) -> Iterator[Statement]:
 
     for name in program.gates:
         if name in stmt_generator_map:
-            warn(f"Gate '{name}' overrides the Jet gate with the same name.")
+            warnings.warn(f"Gate '{name}' overrides the Jet gate with the same name.")
 
         stmt_generator_map[name] = _create_statement_generator_for_composite_gate(
             name=name, stmt_generator_map=stmt_generator_map, gate_signature_map=gate_signature_map
@@ -318,7 +328,7 @@ def _compute_amplitude(
         dtype (type): Data type of the amplitude.
 
     Returns:
-        NumPy number representing the amplitude of the given state.
+        Number: NumPy number representing the amplitude of the given state.
     """
     # Do not modify the original circuit.
     circuit = deepcopy(circuit)
@@ -329,7 +339,141 @@ def _compute_amplitude(
         qudit = Qudit(dim=circuit.dimension, data=data)
         circuit.append_state(qudit, wire_ids=[i])
 
-    # TODO: Find a contraction path and use the TBCC.
+    result = _simulate(circuit=circuit, dtype=dtype)
+    return dtype(result.scalar)
+
+
+def _compute_probabilities(circuit: Circuit, dtype: type = np.complex128) -> np.ndarray:
+    """Computes the probability distribution at the end of a circuit.
+
+    Args:
+        circuit (Circuit): Circuit to produce the probability distribution for.
+        dtype (type): Data type of the probability computation.
+
+    Returns:
+        Array: NumPy array representing the probability of measuring each basis state.
+    """
+    result = _simulate(circuit=circuit, dtype=dtype)
+
+    # Arrange the indices in increasing order of wire ID.
+    state = result.transpose([wire.index for wire in circuit.wires])
+
+    amplitudes = np.array(state.data).flatten()
+    return amplitudes.conj() * amplitudes
+
+
+def _simulate(circuit: Circuit, dtype: type = np.complex128) -> TensorType:
+    """Simulates a circuit using the task-based contractor.
+
+    Args:
+        circuit (Circuit): Circuit to simulate.
+        dtype (type): Data type of the tensor network to contract.
+
+    Returns:
+        Tensor: Result of the simulation.
+    """
     tn = circuit.tensor_network(dtype=dtype)
-    amplitude = tn.contract()
-    return dtype(amplitude.scalar)
+
+    if len(tn.nodes) == 1:
+        # No contractions are necessary for a single tensor.
+        return tn.nodes[0].tensor
+
+    path_info = _find_contraction_path(tn=tn)
+
+    tbc = TaskBasedContractor(dtype=dtype)
+    tbc.add_contraction_tasks(tn=tn, path_info=path_info)
+    tbc.add_deletion_tasks()
+
+    # Warning: The call to contract() below can take a while depending on the
+    # size of the tensor network and the quality of the contraction path.
+    tbc.contract()
+
+    return tbc.results[0]
+
+
+def _find_contraction_path(tn: TensorNetworkType, samples: int = 100) -> PathInfo:
+    """Finds a contraction path for a tensor network. This is done by sampling
+    several random contraction paths and choosing the one which minimizes the
+    total number of FLOPS.
+
+    Args:
+        tn (TensorNetwork): Tensor network to be contracted.
+        samples (int): Number of contraction paths to sample.
+
+    Returns:
+        PathInfo: Contraction path for the given tensor network.
+    """
+    paths = (_sample_contraction_path(tn=tn) for _ in range(samples))
+    return min(paths, key=lambda path: path.total_flops())
+
+
+def _sample_contraction_path(tn: TensorNetworkType) -> PathInfo:
+    """Samples a random contraction path for a tensor network.
+
+    Args:
+        tn (TensorNetwork): Tensor network to be contracted.
+
+    Returns:
+        PathInfo: Contraction path for the given tensor network. Contractions
+        between nodes that share an index are always preferred.
+    """
+    path = []
+
+    # Caching the index-to-edge map improves performance significantly.
+    index_to_edge_map = tn.index_to_edge_map
+
+    # Build an adjacency list using the node IDs in the tensor network.
+    neighbours = {node.id: set() for node in tn.nodes}
+    for node in tn.nodes:
+        for index in node.indices:
+            node_ids = index_to_edge_map[index].node_ids
+            neighbours[node.id].update(node_ids)
+
+        # Nodes are not neighbours with themselves.
+        neighbours[node.id].remove(node.id)
+
+    # Track the set of tensors with no neighbours.
+    isolated = list(sorted(node_id for node_id in neighbours if not neighbours[node_id]))
+
+    for node_id in isolated:
+        del neighbours[node_id]
+
+    # Iteratively contract two adjacent tensors while it is possible to do so.
+    while len(neighbours) > 1:
+        node_id_1 = random.choice(tuple(neighbours))
+        node_id_2 = random.choice(tuple(neighbours[node_id_1]))
+        path.append((node_id_1, node_id_2))
+
+        # Derive the neighbours of the contracted node.
+        node_id_3 = max(*neighbours, *isolated) + 1
+        neighbours[node_id_3] = neighbours[node_id_1] | neighbours[node_id_2]
+        neighbours[node_id_3] -= {node_id_1, node_id_2}
+
+        # Decide whether the contracted node is isolated.
+        if not neighbours[node_id_3]:
+            isolated.append(node_id_3)
+            del neighbours[node_id_3]
+
+        # Replace node_id_1 with node_id_3 and then do the same for node_id_2.
+        for node_id in neighbours[node_id_1]:
+            if node_id != node_id_2:
+                neighbours[node_id].remove(node_id_1)
+                neighbours[node_id].add(node_id_3)
+
+        for node_id in neighbours[node_id_2]:
+            if node_id != node_id_1:
+                neighbours[node_id].remove(node_id_2)
+                neighbours[node_id].add(node_id_3)
+
+        del neighbours[node_id_1]
+        del neighbours[node_id_2]
+
+    if len(isolated) > 1:
+        # The final iteration of the while loop always yields an isolated node.
+        # The ID of this node is always the largest ID in the tensor network.
+        node_id_1 = isolated[-1]
+        for node_id_2 in isolated[:-1]:
+            path.append((node_id_1, node_id_2))
+            node_id_1 += 1
+
+    return PathInfo(tn=tn, path=path)